requirements.txt
flask==3.0.0
python-nmap==0.7.1
python-dotenv==1.0.0
